+++ src/alfasim_sdk/_internal/alfacase/generate_case_description_docstring.py
@@ -69,6 +69,7 @@ CATEGORIES_USED_ON_DESCRIPTION = sorted(
         "power",
         "pressure",
         "productivity index",
+        "roughness",
         "specific energy",
         "specific heat capacity",
         "standard volume per standard volume",